+++ streamlit_app.py
@@ -159,6 +159,9 @@ def filters_sidebar():
             help=_s.DASHBOARD_HELP_SEARCH_LIMIT,
         )
 
+    with st.expander("Filter by panel", expanded=True):
+        st.session_state.filter_panel = st.multiselect("Panel", _s.PANELS)
+
     with st.expander("Filter by unit of assessment", expanded=True):
         st.session_state.filter_uoa = st.multiselect(
             "Unit of assessment", _s.UOA.values()
@@ -211,6 +214,10 @@ def data_section():
 
 
 def filter_data(data: pd.DataFrame) -> Optional[pd.DataFrame]:
+    panels = get_session_filter_panel()
+    if panels:
+        data = data[data[_s.DATA_PANEL].isin(panels)]
+
     uoas = get_session_filter_uoa()
     if uoas:
         data = data[data[_s.DATA_UOA].isin(uoas)]
@@ -246,6 +253,10 @@ def filter_data(data: pd.DataFrame) -> Optional[pd.DataFrame]:
     data = filter_data_by_text_search(data)
 
     return data
+
+
+def get_session_filter_panel() -> list[str]:
+    return st.session_state.get("filter_panel", [])
 
 
 def get_session_filter_uoa() -> list[str]: